It would not let me commit through pushing
When I initially pulled to make sure my branch was up to date it said it would overwrite the gitignore file, which I tried to look up solutions to this but each one warned there risked it deleting the project entirely which I did not want to happen. I could not find a solution and was getting frustrated because rstudio would not let me knit my file either because of an issue with predict(nl...) and "object nl not being found" even though the code worked previously. So after running into all these issues, I decided to simply do a file upload so my final homework is still turned in. Sorry!